Author: Rex Kerr
Fixed some counting bugs related to diffusing molecules not updating their
position enough, and sometimes updating their position to be inside a wall.
This would give incorrect counts if the particle was then destroyed and
counted via count_region_from_scratch because that code couldn't tell which
side it was supposed to be on.

I think I've fixed all the spots where this is an issue, but it still needs
more extensive testing.  The dangerous cases are when you have molecules
being destroyed intermittantly while bouncing around between/through
surfaces that are close together relative to the diffusion step length.