Author: Rex Kerr
Added flags to subvolumes so they can tell when they're at the edge of the
world (and in which direction).

Modified traverse_subvol to use these flags for error checking.

Fixed flag usage in counting of outcome_unimolecular (COUNT_ME is used when
molecules are diffusing; unimolecular reactions never happen *while*
diffusing; the step ends when the unimolecular reaction would occur).