Author: Rex Kerr
Changed formatting of VIZ_OUTPUT section to better match that of the other
sections.  Fixed description of wildcards (they have to be in quotes!).
There is still a mismatch between the description of files and the stated
number of files in some places.

Note: also added a \comment{} macro in the preamble so we can add comments
if we want.  Start a block of LaTeX (with ctrl-L, for example) and type
\comment{Whatever you want as a comment in here} and it won't appear in the
output.

Also added an Authors section a the end, and included myself and Markus in
it.  Has Boris or Tom or anyone else contributed?  Should we make the
section a dual Authors/Changelog section so we have a record within the
document (and not just in CVS messages) of when major changes happened and
who made them?