Author: Rex Kerr
Fixed a counting bug dealing with surface reactions (first reactant doesn't
need to have COUNT_ME set to be counted, since that is only for moving
reactants and 2D reactions always take place with both partners stationary).

Also modified edge_hash to use the jenkins_hash function, and modified that
function to take a length argument also (so it need not work only on
null-terminated strings--the regular hash functions which call it thus call
strlen themselves instead of counting on the jenkins_hash function to call
strlen).  Updated the makefile to add the sym_table.h dependency to
wall_util.c (to see jenkins_hash).

Also added the NULL option for reactions (both in text an as an example),
and fixed some improperly formatted ,, notation (LaTeX was interpreting it
as a lower-double-quote, not as two commas).