Author: Rex Kerr
Changed counting code for diffusing surface molecules.  There's a new
routine called count_moved_grid_mol in count_util that (hopefully) does the
sensible thing when given a grid molecule and the place where it will move
to.  The calls to count_region_from_scratch in the diffuse_2D code have been
replaced with this new function.

In order to get counting to work more efficiently with large numbers of
regions, I also added code to sort the region lists as added to walls by the
memory address of the region.  This policy means that output order of
TRIGGER statements are not necessarily identical from platform to platform.
It may be smarter to add a global index for regions and sort off that
instead using void_list_sort_by.

Also, the size of buffers for triggered functions was kind of weird.  I
fixed it so that the buffer size for triggers is not set to a small value if
only a small number of COUNT statements would be output--it either is the
default value (currently #defined as 10000) or the user-supplied value.

Finally, I added flags and changed the output format, since there are three
different cases of triggering.

Reactions have format
  it_time  exact_time  X  Y  Z  [name]
since if a reaction is triggered, there is always one of it.

Triggering off of hits gives the format
  it_time  exact_time  X  Y  Z  orient  [name]
where +1 means the front face was hit/crossed and -1 means the back face
was.

Triggering when the number of molecules changes gives
  it_time  exact_time  X  Y  Z  orient  number  [name]
where number==+1 means that the molecule was added and -1 means it was
removed.  Orientation is 0 for volume molecules rather than leaving out the
field.