Author: Rex Kerr
Added new microscopic reversibility routines to match angles as well as
distances for unbinding from surfaces, and to match the binding distribution
when performing unbinding reactions in the volume.

My very preliminary tests indicate that the slowdown in these extra routines
is quite small, and should probably be the default settings (in fact, if
they work as well as the preliminary tests indicate, there probably should
be no other option!).  However, right now the default is to not perform
microscopic reversibility so old simulations aren't broken.

To change the setting, use
  MICROSCOPIC_REVERSIBILITY = ON
     or                     = OFF
     or                     = SURFACE_ONLY
     or                     = VOLUME_ONLY
depending on what you want to be reversible.

This is also explained (very briefly) in the QRG.  The mathematics behind it
is explained in the MCell3 methods paper.