Tom Bartol:
Modify output format of chkpt file to include the seed_seq used to initialize
the rng.  This is used to detect whether it is necessary to reinitialize the rng
between chkpt runs or whether it is O.K. to continue with the rng state retrieved
from the chkpt file.  This is done as follows:

Compare seed_seq in chkpt file with seed_seq from command line.  If seeds
match, continue with sequence from chkpt file, otherwise, reinitialize rng to
beginning of new seed sequence.  Note that in either case we need to fread the
rng data from the chkpt file to advance to the next cmd in the chkpt file.