Quick fix for floating point exception from viz output module when loading a checkpoint under certain circumstances.  The error occurs when a checkpoint file is loaded and checkpointing is not enabled in the subsequent run.  The viz output code has the following check:
  /* here is the check to prevent writing twice the same info
     - at the end of one checkpoint and the beginning of the next checkpoint.
   */
  if((world->chkpt_flag) && (world->start_time > 0)){
    if (world->it_time == world->start_time){
      if(viz_iteration % (world->chkpt_iterations) == 0){
        return 0;
      }
    }
  }

Essentially, what this code is trying to guard against is the following: If
MCell is set to produce viz output every m iterations and checkpoint every n
iterations, every (m*n/gcd(m, n)) iterations, it will be both producing viz
output and checkpointing.  If an attempt is made to resume that checkpoint,
the viz output will produce another output file immediately upon starting up,
giving a duplicate data file.  So, in the case that a checkpoint is loaded, on
the very first iteration only, if the current viz iteration is a multiple of
the checkpoint iteration count, it skips the output.

Unfortunately, the checkpoint iteration count can change after restarting from
a checkpoint, and can even be 0 if checkpointing has not been enabled.  As a
result, I don't think the check is correct, and it can lead to the
divide-by-zero which has been reported earlier.

There are a few solutions:

  1. remove the check and just live with an occasional duplicate data file

  2. store enough info in the checkpoint file to determine whether viz output
     was produced after the last iteration before the checkpoint and use that
     instead

  3. transition the viz output over to using a scheduler like the other parts
     of the code.  after talking to Tom, it sounds like we could probably use
     the scheduler's mechanism for dealing with checkpoints to avoid the
     duplicate output file problem

For now, however, I've simply modified the check to make sure we aren't
dividing by zero.  This means that the two most obvious scenarios will work
quasi-correctly:

    1. using the same checkpoint interval (or a multiple of the same
       checkpoint interval) after the checkpoint continues to work completely
       correctly as it does now

    2. not setting checkpoint interval (i.e. disabling checkpointing) works
       correctly, with the caveat that you might get an extra output file
       occasionally