Fixing a number of minor memory leaks and one significant memory leak.
The significant memory leak is in the exact disk calculation.  exd_vertex
objects are maintained in (approximately) a linked list throughout the exact
disk calculation, and thrown away when the calculation finishes.
Unfortunately, not all items were in the linked list -- some were attached to
the data structure via a 'span' pointer.  Any objects which ended up in the
'span' linked list for a given vertex would not be properly freed.  I added
logic at the end to flatten it all into a proper linked list.

I also added a couple of lines to bulk-initialize the 'world' and 'mdlpvp'
objects when they are allocated, both to avoid being bitten by uninitialized
fields at some later date, and to quiet valgrind's complaints about the 20 or
30 fields in these structures which are used uninitialized.  (Looking over all
of the fields about which it is complaining, 0/NULL/0.0 seem to be the right
initialization value.)