Fixed a crash in the viz_output code.  If a run was resumed from a checkpoint and checkpointing was not enabled after the resume, the allocated buffers for viz data could be too small, leading to buffer overruns, which could lead both to mysterious results, and to crashes.
It seems that it might be useful for chkpt_flag to be split into two flags,
maybe 'chkpt_restored', and 'chkpt_enabled'.  Some places in the code seem to
use chkpt_flag to mean that checkpointing is enabled.  chkpt_flag is also a
slightly opaque name.

Also, cleaned up the error message for a missing viz_data directory.  The old
was somewhat obscure.