Fixes for a few more incorrect units.  Polygon data was being stored both in world and internal coordinates, with the world coordinates used for most purposes, but the internal coordinates used for computing bounding boxes. This was causing a few odd behaviors.  First, since the world bbox was in the wrong coordinates, the partitions were being created incorrectly in some cases, and a message about "rescaling" would appear.
Because the partitions were created slightly wrong, we had an example where
the walls of a box were contained in the partition which extends off to
infinity.  This caused the waypoints to be a long way off, and the raytracing
from the waypoint to the point of molecule creation (for instance, in
concentration clamp) was not always accurate enough.  The result of this is
that the world counter and the enclosed volume counter for this molecule were
becoming out-of-sync (i.e. the box contained molecules which were not included
in the count).

I'm not sure whether it's possible now to cause actual geometry to appear
in one of these "infinite" partitions.  If not, it may be worth adding some
more logic to set_partitions to make sure this doesn't happen if the user
picks particularly bad geometry.  I'll look into this, and if it seems
possible, I'll add the required logic.