Diffusion bug.  In some cases, it was possible for a molecule to diffuse into coordinates owned by another subvolume but without detecting a collision with the sv boundary.  This could lead to all sorts of mayhem, including particles travelling through walls.
In particular, the bug occurs if all of the following are satisfied:

    1. The velocity vector consists of 3 components, exactly two of which are
       non-zero

    2. The non-zero components of the velocity vector have opposite signs

    3. In the two directions which have non-zero velocity components, one
       direction has to travel for more than one timestep to hit the partition
       boundary, and the other has to travel for less than a timestep.

The problem arises because of a test of the following form:

    tx = dx*vz
    tz = dz*vx
    if (tx < tz) ...

The derivation for this is the following.  dx and dz are the distances to the
nearest x and z partition resp. in the direction of travel.  We want to compare
the times at which we reach each of these:

    dx / vx      versus     dz / vz

and take the smaller of the times.  Now if:

    dx/vx < dz/vz

we can multiply both sides by vx*vz to eliminate the division.  This only
works, however, if vx and vz have the same sign.  If the signs differ, so does
the polarity of the comparison.  I added code to take the absolute value of
these tx, ty, tz before doing the comparison.