Another rare bug which allows molecules to leak through surfaces.  This is in almost the exact same place as the last bug, and appears to have been a copy-and-paste error.  I can't believe I didn't spot it the last time through this function!  Essentially, when a molecule was diffusing in the x-y plane (i.e. the z-component of the direction was 0), the test for which sv bound would be hit first was:
        tx=dx*v->z;
        ty=v->y*dy;
        ..
        if (tx<ty)
        ..

but it should've been:

        tx=dx*v->y;
        ty=v->x*dy;
        ..
        if (tx<ty)
        ..

(i.e. v->x, v->y, and v->z were permuted).  The end result is that a molecule
could get into a subvolume to which its coordinates did not belong, just as
with the most recent previous bugfix I checked in.  (I wonder if there's a good
place to add some sanity checking so that we catch molecules which don't move
to the proper subvolume when they diffuse.  This might be worth looking into,
just so we can quickly track down problems of this sort if any more occur.)