As per the suggestion of Rex and Boris, migrating the general utility functions from viz_output.c to util.c.
There are getting to be a number of filesystem-related utilities in the util.c
file, many of which write to world->err_file.  It may be about time to split
them off into file_util.c or a similarly named file.  Actually, it wouldn't be
a bad idea to work up a set of standardized file utilities and use them
consistently, to ensure that error handling and reporting are uniform
throughout the application.