Author: Rex Kerr
Added the ability for closed regions to clip surfaces upon which you put
moelcules.  The code appears to work (with simple testing), and I fixed a
bug with multiple overlapping surface region releases along the way.

In order to do this, use
  SHAPE = surface_region % volume_region
where "surface region" can actually be any combination of surface regions,
and "volume region" can actually be any combination of volume regions.  The
code doesn't yet check to make sure the volume regions are all closed; it
probably ought to.