Update to handle coincident surfaces.  There was a bug in the handling of coincident surfaces.  Essentially, if there were surfaces which were transparent to a particular molecule coincident with surfaces which were reflective to the same molecule, we would skip reporting the outbound crossing with the transparent surface, because we reflected rather than crossing; however, upon reflection, we WOULD count the inbound crossing.  Note that this effect is dependent upon the ordering of the walls in the subvolume's wall list, as it only applies to walls which appear before the reflective wall in the wall list.
One effect of this bug is that counts could go negative or very positive, if
the transparent surface in a coincident surface pair was an object inside of
which we were counting a molecule.  This occurred because the molecule, when
reflecting off of the surface, could be considered to enter the volume even
though it never entered, or enter the volume though it never left.