Made REACTION_DATA_OUTPUT require an output timing specification.
At present, it will give a fairly non-descript parse error if no STEP (or other
output timer) is specified.  We can make it more specific if we need to later.

Note that immediately prior to this revision is a version which simply restores
the old MCell 3 behavior.  That way, we can go back to it easily just by
reverting this single commit if we change our minds.