Fix for a bug related to volume output when restoring from checkpoints.
If volume output were turned on and a checkpoint is loaded from a time after
the last scheduled volume output, the first iteration would take a VERY LONG
TIME, restricting the maximum time step to the tiniest fraction of an
iteration.  The problem is that it would schedule the volume output anyway, but
it wouldn't process it until the end of the iteration, and until then it would
show up as a barrier preventing the adaptive timestep from being larger than
epsilon.