Fix for the -e option to main.py: tests should now be properly excluded.
I've updated main.py to fix the bug Boris encountered.  Tests were not properly
being excluded due to a logical error in the setup of the run tests list.
We have to expand the list before we can remove the excluded tests from it.
Instead, the list was being expanded AFTER exclusion.  Oops!