Added more checks for place_waypoints_flag.
There were a few places yet which were not properly checking for
place_waypoints_flag, and were, thus, triggering attempts to handle volume
molecule and reaction counting on regions, even when no waypoints were
available.
  
To be safe, a catch-all has been added to the appropriate section of
count_region_from_scratch to avoid this issue.  Note that this may obsolete
many of the specific checks of this flag throughout the code, although some of
them may convey performance advantages if they allow quickly skipping blocks of
unnecessary code.