Fix for counting of ENCLOSED surface molecules.
This is a two-fold fix.

First, we were updating the n_at counter instead of the n_enclosed counter.
This caused the count for the enclosed surface molecule to never be updated,
resulting in a count which remained constant.

The second bug, which was masked by the first bug, was in the function
next_subvol, which is used to select the "next" subvolume along a diffusion
ray.  If one of the movement vector coordinates was precisely zero, it could
sometimes choose the wrong next subvolume.  (In particular, the bug was
triggered if the Z coordinate of the movement vector was zero, and the molecule
should have hit the X partition first.  There may have been other problematic
cases.)  The symptom of the second bug, at least in the context of Markus' test
case, is that the counts in the left and right boxes, while continuing to sum
to 50, would individually become negative or exceed 50.

While this latter bug could have caused other problems, it is extremely
unlikely, as the relevant code is mostly used for volume diffusion.  It's far
easier to replicate the bug in surface diffusion, in which movement may be
easily constrained to an axis-aligned plane, or in the case of an axis-aligned
cuboid, your choice of 12 different axis-aligned triangles.

Regression test checks for both failure modes.