Fixed bug with delayed counting.
Instead of updating the counts, we were overwriting them with arbitrary and
incorrect values.  This may lead to unexpected results.

In particular, integer counts were being overwritten by the delta which was to
be applied to that particular count on behalf of one of the worker threads,
and double counts were being half overwritten by an integer value cast from
the pointer to the double value containing the delta.