Add several long-forgotten documents related to reactions and our spec on the output log.  These are useful for understanding MCellv3, of course, but will also be very helpful in refactoring and redesign in MCellv4/libMCell.
Pay special attention to rdl.0.2.0.pdf.  This documents the early thinking of
Rex, Joel, and myself for a more comprehensive reaction description language
for reactions and complex reactions in MCell.  Already in 2004 I think we came
close (but no cigar) to BNGL.  This may be useful for our thinking as we merge
and further generalize BNGL with MCell RDL so that we can have reactions and
self-assembly of macromolecular complexes in space (at a much coarser scale
than required of fine-grained molecular dynamics).