Disabled "poor" user warning.
For the time being I disabled the message that would tell users

"Warning: Probability less than EPS_C for reaction between ..."

since it is not at all clear what a user is supposed to do when
probabilities are smaller than EPS_C? The underlying issue really
seems more a code design issue, namely to what numerical precision
can we guarantee accurate reactions and what to do if this condition
is violated. I have the feeling that our current treatment via EPS_C
is not consistent and also not very portable but I may be wrong.