Added code for custom binary reaction data writer.
This branch contains code for a custom binary reaction data writer.
Below, I will provide a brief summary of the purpose and capabilities
of the code - more details can be found in additional documentation
which I will write up shortly.

In the process of adding the binary writer code I've also rewritten
some of the infrastructure for the existing reaction data code which
is now somewhat cleanly separated from the rest of MCell. Thus, adding
additional output writers (such as HDF5, SQL) should now be fairly
straightforward (I think). This is not yet a clean data-out API but
definitely much closer to what we currently have in the main line.

Back to the binary reaction output data writer:

In a nutshell, MCell currently outputs each COUNT statement inside
reaction data blocks as separate ascii files. This can become
cumbersome for systems with large numbers of COUNT statements
(such as our active zone models which produce thousands of count
files per single seed). The binary reaction data writer instead
outputs all COUNT statements within a single reaction data block to
a single streamed binary file (the exact format of which will be
explained in a separate document).

In addition, the binary writer can compress the data on-the-fly
enabling a very compact and thus IO friendly way to write data to disk.
Depending on the model, compression can be very good. E.g., the reaction
data for our active zone model compresses by factors of 10^2 to 10^3
(since many of the COUNTS are zero for the duration of the run).
Based on some benchmarking the cost for on-the-fly compression is very
small and more than compensated for by the very nice IO characteristics
of the compact binary output files especially when hundreds of MCell
processes working on different seeds are writing to the same file system.

The new keywords for the reaction data output block which control
the binary reaction output are

BINARY_OUTPUT = <TRUE|FALSE>
BINARY_OUTPUT_COMPRESSION_LEVEL = <1..9>
BINARY_OUTPUT_COMPRESSION_TYPE = <GZIP|BZIP2>
BINARY_OUTPUT_DIRECTORY = "<string>"
BINARY_OUTPUT_FILENAME = "<string>"

If

BINARY_OUTPUT = FALSE

which is the default MCell uses the traditional ascii output
format. Binary reaction data output is requested on a per
reaction block basis. I.e., each block can be output as either
binary or ascii data depending on need.

The main drawback of a custom binary output format is of course
the need to have a separate tool to extract the data. For this purpose
I've written a small C++ library and some command line tools that
make this fairly easy and painless. I will commit these shortly
and provide more details.

IMPORTANT NOTES/TODO:
---------------------

- Currently, the binary reaction output writer does *not* support
  TRIGGER statements only COUNT statements. Thus, if a mix of
  COUNT and TRIGGER statements is required an easy workaround is
  to group all TRIGGER statements into a separate reaction data
  block which can then be output in ascii while writing the COUNT
  statements in binary format.

- The binary reaction output writer currently does not support
  checkpointing and subsequent checkpoints will overwrite previous
  files instead of appending without additional user intervention
  (like moving files out of the way in between checkpoints).
  The binary output format does in principle allow appending so proper
  checkpointing could be implemented in the future.

- The on the fly compression adds external dependencies on libz and
  libbz2. The configure file will check for their presence. In
  principle, compression could be made optional. Something to consider
  even though at least libz should be present on most systems out
  there.

- This code should be considered beta. While it has been heavily
  used for our active zone model it hasn't seen a lot of other MCell
  models so some bugs may still be lurking.

- The binary output format is currently not endianism aware. This
  needs to be fixed. It should be fairly straightforward using
  MCell's existing facilities for checkpointing.