Give up if we can't move point off edge
In closest_interior_point, we check to see if a point is on an edge. If
it is, we move it toward the centroid. However, in certain cases, the
movement is so small that the simulation is effectively stuck in the
loop.

The fix in this commit simply bails out if we can't move off the edge
after a set number of attempts. The testsuite passes, and this fix poses
no problems that I'm aware of. In fact, it's a non-issue in most
cases since one movement is usually sufficient to move off the edge.

We should consider revisiting this function in the future, as there
might be more efficient ways of handling this problem. For instance, as
an alternative solution, I attempted to move the point by a set
distinguishable amount (instead of incrementing by some fractional
value), but that failed one of the macromolecule tests
(03-macro_surface_listrelease.mdl).