Made testsuite error message more useful
When certain tests in the testsuite fail, it is because a value is out
of tolerance. It can be useful to know how much it's out of tolerance,
but you have to check the python file to see what the value and
tolerance actually is. This commit puts that information in the error
message itself.