Fixed radial directions/subdivisions bug
If radial directions or subdivisions was set to zero, it would result in
a divide by zero error. The error checking for radial directions did not
include a check to see if it was a positive value. Such a check existed
for radial subdivisions, but the radial directions check was being
called instead. Both of these problems have been fixed.

Some refactoring was also done.