Added API functions to set iterations and timestep
Up until now, the API function calls have only returned success or
failure. The API function mcell_set_time_step uses unique error codes,
which will give users more detailed information if they want it.
Perhaps, there's a more elegant way to handle this, but this is a simple
case in which to test it.