More work on creating geometry with the API.
A portion of this  commit involves creating the default region (ALL) and
everything entailed with that.

There is more "redundant" code in this commit than in previous ones
related to the creation of geometry. The reason for this redundancy is
that the API and parser need to do very similar things, but the API
needs to be quiet. For a large function with numerous fail points (and
thus print statements), this is a problem. Some of these problems can
and will be fixed with the use of error codes. Another issue is that
some of the things which are errors in the parser should perhaps not be
an error at all in the API.