Fixed how the printing of species info is handled.
The final step of parsing molecules/species definitions is printing
summary information about them. This was directly interacting with the
internal "world" simulation state instead of using information received
via the API. This has been resolved.

Unfortunately, this meant re-incorporating sections of code which were
removed during the last commit. It might be worthwhile investigating
this in the future to see if some of the redundant sections can be
removed.