Refactored release site code. Updated reporting.
The following changes were made in refactoring the code:

- Removed some old code.
- Commented out some old functions.
- Moved declarations closer to where they are used/assigned.
  (Presumably it was done this way b/c of C89 standard.)
- Broke up big functions.
- Made some overused copy-paste code into a function

There are still some fairly large functions with a bit of redundancy in
them that could be refactored some more.

----------

Previously, we reported release events like this:

Releasing X molecules A ...  Released Y A from "World.Thing" at
iteration 0.

This is usually redundant at best and confusing in the case when X and Y
are not the same thing. This can happen if you are releasing a certain
concentration of volume molecules into an object that is a different
volume than the bounding box (i.e. almost everything), because X is the
number of molecules that would be released IF the molecules were being
released into the bounding box and Y is what is actually released into
the object. These numbers can be drastically different, and the user has
no idea why X is what it is. As such, all such "Releasing" statements
were removed, and now we only report on what is actually released.
Also, this change does not impact the warning that is reported if you
try to release more surface molecules than allowed by the
SURFACE_GRID_DENSITY.