Consolidated more copy-paste code.
Like the previous commit, this consolidates code related to the release
and placement of a single surface molecule. The function created in this
commit is used to place molecules via surface classes and the function
from the previous commit places molecules via release site objects.
There is quite a bit of overlap to both functions and they should be
consolidated if at all possible.