Removed code related to region inclusions.
NOTE: If we want to bring back this code just revert this commit.
However, if you do so you need to provide documentation on what this
feature is supposed to do, provide a rationale for the actual code and
also a test case. It is possible that the code as is does not work
correctly so additional testing and bugfixing will be required
before re-incorporating the removed code.

Rationale: The region inclusion feature was undocumented and nobody
on the current team really knew what it was supposed to do so it
seemed prudent to remove it for the time being.