Updated QRG release number and title
The release number (3.2.1) was also being used for the version number
(3.2). The html title was using the first section heading (Running
MCell3) in the document, which might give the wrong impression about the
intent and contents of the document. The current solution doesn't seem
ideal, but it's probably good enough for now.