Define properties for surface classes via API.
Reflective, absorptive, and transparent surface class properties can be
defined via the API now. The concentration clamp property still needs to
be added. The wording of some functions (e.g.
mcell_add_surf_class_properties) may need to be adjusted so they are
consistent with the QRG, which considers any of the commands (including
MOLECULE_DENSITY/NUMBER) defined in a surface class to be a "surface
property".