Change resize factor when shrinking pointer_hash.
Previously, the code would shrink the pointer_hash to the
number of items during removal of items from the hash. This
seems to small since any addition of new items will cause
an immediate resize to 2*(number of items). Thus, I changed
the code to shrink the pointer hash to 2*(number of items).