Reincorporated existing dynamic mesh code.
This is an initial commit to restart work on dynamic meshes. This brings
some of Boris's old work in line with the current state of the git
repository.

This commit adds in functions to find enclosing meshes and move volume
molecules as required, but they are not being used in this commit yet.
Also, the changes to insert surface molecules still need to be added.

Not all of Boris's changes will be added, since his method relied on
checkpointing, and the new approach will not. For instance, it may
instead work in a manner similar to variable rate constants where a file
contains two columns, in this case, one being time and the other being a
filename (those files containing the meshes to be used at the specified
time).