Bugfix: Initializing symbol table for counting
This fixes a problem that was introduced in a recent refactoring.  With
that said, it's not really clear what the the point of the
counter_by_name symbol table is. Aside from the functions touched in
this commit, it's only used by a function that isn't called during a
normal simulation (get_counter_trigger_column). As such, the bug didn't
cause any serious breakage.