Replace geometry and place volume molecule.
This commit allows geometry to be changed without checkpointing.
Eventually, this will be a scheduled event, but for now it's just hard
coded for testing purposes. First, we save all of the volume molecules
currently in the scheduler. Then the old geometry is trashed (along with
a lot of dependencies), the new geometry is parsed, and initialized.
Finally, we insert all molecules into the world using Boris's existing
code.