Added new keyword, DYNAMIC_GEOMETRY, to parser.
Here is an example of how this keyword would be used in an MDL:

  ITERATIONS = 10
  TIME_STEP = 1E-6
  DYNAMIC_GEOMETRY = "mesh_changes.txt"
  INCLUDE_FILE = "my_geom.1.mdl"

  INSTANTIATE world OBJECT {
    Cube OBJECT Cube {}
  }

Also, let's suppose my_geom.1.mdl looks like this:

  Cube POLYGON_LIST
  {
    VERTEX_LIST
    { ...  }
    ELEMENT_CONNECTIONS
    { ...  }
  }

The file referenced by the DYNAMIC_GEOMETRY command, mesh_changes.txt,
would contain a list of times and MDL file names like this:

2e-6 my_geom.2.mdl
3e-6 my_geom.3.mdl
4e-6 my_geom.4.mdl
...

The first columns refers to the time at which the geometry will be
changed. The second columna refers to a file that contains all the new
polygon objects. This is very similar to how variable rate constant
files work, except the second column is an MDL name instead of a rate
constant.

This means that the meshes originally instantiated at the start of the
simulation in my_geom.1.mdl would be replaced at 2e-6 by those contained
in my_geom.2.mdl. And those in my_geom.2.mdl would be replaced at 3e-6
by my_geom.3.mdl and so on.

Every polygon object that was instantiated in the original MDL should
exist in the geometry files referenced. The polygon objects should also
be reinstantiated in these files. Here's what my_geom.2.mdl might look
like:

  Cube POLYGON_LIST
  {
    VERTEX_LIST
    { ...  }
    ELEMENT_CONNECTIONS
    { ...  }
  }

  INSTANTIATE world OBJECT {
    Cube OBJECT Cube {}
  }

These limitations may change in the future, but it is currently
necessary.