Better placement rules for surface molecules.
This is a slight modification of Boris' original code to place surface
molecules. When placing a surface molecule after a geometry change,
MCell will try to place it on a surface where the object and region
names match that of whatever surface it was previously on. This prevents
problems where a surface molecule might unrealistically hop to a surface
on a completely different object or region.