Add data struct w/ species-object transp. info.
During simulation initialization, a data structure is created that
stores what species are transparent to what objects. This data structure
uses a pointer hash with species names as keys and values are a linked
list of pointers with type object_transparency. The object_transparency
struct contains the object's name and whether the species can go
in-to-out and/or out-to-in.

This data structure is primarily intended for use with dynamic
geometries. The current implementation of dynamic geometries ignores
geometries that aren't watertight for purposes of placement. Given this
constraint, we treat partial transparencies as fully transparent.